Add `--trace-dispatch`
Similar to `--trace-compile`, emit the `precompile` statement for a method
once, but only when it is dynamically dispatched.

For this, we add a `dispatched` field into `jl_method_instance_t` that is
initialized to 0, and when the method is dispatched, it is set to 1 and the
precompile statement is emitted (once).